requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
appnope==0.1.4
argcomplete==3.6.0
asttokens==3.0.0
attrs==25.3.0
autogen-agentchat==0.4.9.2
autogen-core==0.4.9.2
autogen-ext==0.4.9.2
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
azure-identity==1.21.0
beautifulsoup4==4.13.3
blessed==1.20.0
boto3==1.37.10
botocore==1.37.10
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.4
cohere==5.14.0
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
editor==1.6.6
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.11
fastavro==1.10.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
google-auth==2.38.0
googlesearch-python==1.3.0
greenlet==3.1.1
griffe==1.6.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.5.0
inquirer==3.4.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jiter==0.8.2
jmespath==1.0.1
json_repair==0.39.1
jsonpath-python==1.0.6
jsonref==1.1.0
jupyter_client==8.6.3
jupyter_core==5.7.2
logfire-api==3.7.1
lxml==5.3.1
magentic-one-cli==0.2.3
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.5.1
msal==1.32.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
olefile==0.47
openai==1.65.3
openpyxl==3.1.5
opentelemetry-api==1.30.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.3
pdfminer.six==20240706
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.50.0
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.36
pydantic-ai-slim==0.0.36
pydantic-graph==0.0.36
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-googlesearch==1.1.1
python-pptx==1.0.2
pytz==2025.1
pyuseragents==1.0.5
PyYAML==6.0.2
pyzmq==26.3.0
ratelimit==2.2.1
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
runs==1.2.2
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
stack-data==0.6.3
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xmod==1.8.1
yarl==1.18.3
youtube-transcript-api==1.0.1
zipp==3.21.0
